refactor(userspace): extract NAT wire types to protocol_nat.go
Pure code-motion split of the pkg/dataplane/userspace/protocol.go
wire-protocol god file (#4839). This first increment lifts the
source/static/destination NAT, NAT64 and NPTv6 rule snapshots plus the
NAT match helper wires out of protocol.go into a cohesive sibling
protocol_nat.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change): NatPortRangeWire,
NatAppTermWire, SourceNATRuleSnapshot, StaticNATRuleSnapshot,
DestinationNATRuleSnapshot, NAT64RuleSnapshot, Nptv6RuleSnapshot.

Behavior is byte-identical; the Rust userspace-dp helper deserializes
these structs, so every json tag is preserved unchanged. Build of
pkg/dataplane/userspace passes; gofmt clean.
